apps/iot-load-tester/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
locust>=2.20.0
paho-mqtt>=2.0.0
python-qpid-proton>=0.39.0
httpx>=0.26.0
prometheus-client>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
